+++ src/include/smash/input_keys.h
@@ -607,6 +607,17 @@ class Key {
  * setup. Multiple events per file are supported. In the following, the input
  * keys are listed with a short description, an example is given and some
  * information about the input particle files is provided.
+ *
+ * \attention
+ * In `List` modus, the provided list of particles has to match information
+ * contained in the particles file (either the SMASH default one or that
+ * provided via the `-p` option), when appropriate. In particular, the mass of
+ * stable particles has to match that of the particles file. In case of a
+ * mismatch, the latter is used (modifying its energy to put the particle back
+ * on shell) and the user warned. Furthermore, all particles have to be on their
+ * mass shell. If not, their energy is adjusted and the user warned. Note that
+ * this type of warning is given only once and <b>it is user responsibility to
+ * ensure that this is a desired behaviour</b>.
  */
 
 /*!\Userguide
@@ -5777,7 +5788,7 @@ General:
              Jet_Momentum: 100.0
 \endverbatim
  *
- * \note
+ * \note\anchor modi_box_usage_remark
  * The box modus is most useful for infinite matter simulations with thermal and
  * chemical equilibration and detailed balance. Detailed balance can however not
  * be conserved if 3-body decays (or higher) are performed. To yield useful