+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 * The box modus can only be used with the fixed time step mode
 * Avoid integer overflow to silently happen in grid construction
 * Correct tau lepton mass in particles list files
+* Fix code behavior in list modi setup to be coherent with given warnings in case of input discrepancies
 
 ### Changed
 * ⚠️ The `master` branch has been renamed to `main`